refactor: extract api key DB lookup to findApiKeyByHash helper
- Add findApiKeyByHash() to helpers/apiKeys.ts
- authenticateHttpApiKey now uses the helper instead of direct DB query
- Domain logic (revoked, expired checks) stays in the auth util
